+++ src/main/java/io/kontur/eventapi/entity/Feed.java
@@ -15,4 +15,5 @@ public class Feed {
     private List<String> roles;
     private List<String> enrichment;
     private String enrichmentRequest;
+    private List<String> enrichmentPostProcessors;
 }